requirements.txt
annotated-types==0.7.0
asgiref==3.7.0
asttokens==2.4.1
async-timeout==4.0.3
cachetools==5.3.3
certifi==2024.2.2
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dotenv==0.0.5
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.3
fonttools==4.53.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
ipython==8.26.0
jedi==0.19.1
kiwisolver==1.4.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
msgpack==1.0.8
numpy==2.0.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pipenv==2024.0.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2021.1
redis==5.0.5
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.4.1
stack-data==0.6.3
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.2
wcwidth==0.2.13
